Add files via upload
This is a loan Assistant app that calculates the figures involved in any loan taken by an individual. The minor modifications include the user interface change, code has been rewritten using Java 8 and 11 features.
This project will be updated later to offer records of assessed loans report, a delayed animated result display and a time stamp on when the loan should be completely paid.